reorganized reactor code and moved some of its support code to the main jmg library
